+++ example/bin/example.dart
@@ -5,12 +5,12 @@ import 'package:stack_trace/stack_trace.dart';
 import 'package:team_logger/team_logger.dart';
 
 final theme = LogMainTheme.defaultActiveTheme.copyWith(
-  // hiddenStyle: LogMainTheme.defaultActiveTheme.hiddenStyle.resetInvisible,
-);
+    // hiddenStyle: LogMainTheme.defaultActiveTheme.hiddenStyle.resetInvisible,
+    );
 final inactiveTheme = LogMainTheme.defaultInactiveTheme.copyWith(
-  // hiddenStyle: LogMainTheme.defaultInactiveTheme.hiddenStyle.resetInvisible,
-  // minLevel: LogLevels.debug,
-);
+    // hiddenStyle: LogMainTheme.defaultInactiveTheme.hiddenStyle.resetInvisible,
+    // minLevel: LogLevels.debug,
+    );
 
 final logStorage = LogStorage(maxCount: 100);
 
@@ -24,7 +24,7 @@ final log = Logger('app')
       inactiveTheme: inactiveTheme,
       isLogActive: (log) => true,
       // activeLevel: LogLevels.error,
-      // activeLoggers: {'events'},
+      // activeNamespaces: {'events'},
       // activeTraceGroups: {'feature'},
       // activeTags: {'response'},
       // isLogActive: (log) => log.hasData,