+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.0
+
+- Update README.
+- [breaking changes] Rename `activeLoggers` to `activeNamespaces`.
+- [breaking changes] Rename `activeLevel` to `activeMinLevel`.
+- [breaking changes] Remove `AnsiPair`.
+- Add `activeLevels`.
+- Fix minor bugs.
+
 ## 0.2.0-0.2.2
 
 - Publish.